linked up pages so we can click through most
