+++ public/js/land.js
@@ -1,0 +1,7 @@
+$(".logInButton").on("click", function(){
+    window.location.href = "/login"
+});
+
+$(".createAccountButton").on("click", function(){
+    window.location.href = "/createaccount"
+});
