+++ routes/apiRoutes.js
@@ -1,10 +1,12 @@
 var db = require("../models");
+var path = require("path");
 
 module.exports = function(app) {
   // Get all trails
-  app.get("/api/trails", function(req, res) {
+  app.get("/api/trails/", function(req, res) {
     db.Hike.findAll({}).then(function(trailInfo) {
       res.render("trail-card", trailInfo);
+      // res.json(trailInfo);
     });
   });
 
@@ -15,9 +17,12 @@ module.exports = function(app) {
     });
   });
 
-  // Create a new example
+  // Create a new user
   app.post("/api/user", function(req, res) {
-    db.User.create(req.body).then(function(newItem) {
+    console.log(req.body)
+    db.User.create({
+      email: req.body.email
+    }).then(function() {
       res.redirect("/");
     });
   });
